refacto StepperModal
