+++ components/modals/StepperModal.js
@@ -1,9 +1,10 @@
 /* global UIkit */
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
 import PropTypes from 'prop-types';
 
 import { CloseButtonNoSSR } from '../utils/CloseButton';
 import HeaderModal from './HeaderModal';
+import { ModalContext } from '../store/ModalProvider';
 
 /**
  * Ce composant fournit une modal à contenu variable selon l'index ou l'on s'y trouve
@@ -11,31 +12,39 @@ import HeaderModal from './HeaderModal';
  * cela permet de gérer le flux/ la modale depuis ses composant internes
  */
 const StepperModal = ({ composers, title, id, resetForm }) => {
-  const [index, setIndex] = useState(0);
+  const {
+    close,
+    setClose,
+    next,
+    previous,
+    index,
+  } = useContext(ModalContext);
+  // const [index, setIndex] = useState(0);
   const [wrappedComponents, setWrappedComponents] = useState();
-  const close = () => {
-    resetForm();
-    UIkit.modal(`#${id}`).hide();
-    // TODO: Probleme car il est possible que la modale se ferme par un moyen autre qu'ici (uk-close-icon~)
-    setIndex(0);
-  };
-  const next = () => {
-    setIndex(index + 1);
-  };
-  const previous = () => {
-    setIndex(index - 1);
-  };
+  // const close = () => {
+  //   resetForm();
+  //   UIkit.modal(`#${id}`).hide();
+  //   // TODO: Probleme car il est possible que la modale se ferme par un moyen autre qu'ici (uk-close-icon~)
+  //   setIndex(0);
+  // };
+  // const next = () => {
+  //   setIndex(index + 1);
+  // };
+  // const previous = () => {
+  //   setIndex(index - 1);
+  // };
   useEffect(() => {
     if (composers) {
       setWrappedComponents(
-        composers.map((composer) => composer(close, next, previous))
+        composers.map((composer) => composer(() => (setClose(close)), next, previous))
       );
     }
   }, [composers]);
+
   return (
     <div id={id} className="uk-flex-top" data-uk-modal="bg-close:false">
       <div className="uk-modal-dialog uk-margin-auto-vertical uk-width-2-3@m uk-width-1-2@l">
-        <CloseButtonNoSSR className="uk-modal-close-default" onClick={resetForm}/>
+        <CloseButtonNoSSR className="uk-modal-close-default" onClick={resetForm} />
         <div className="uk-modal-body uk-padding-large">
           <HeaderModal>{title}</HeaderModal>
           {wrappedComponents && wrappedComponents[index]}
@@ -56,7 +65,7 @@ StepperModal.propTypes = {
 };
 
 StepperModal.defaultProps = {
-  resetForm: () => {}
+  resetForm: () => { }
 };
 
 export default StepperModal;
